RFC Move tests to test_common
